requirements.txt
Flask
pytest
cryptography
requests
PyJWT